dashboard/requirements.txt
flask
flask-sqlalchemy
python-dateutil
psycopg2-binary
gunicorn
python-dotenv
requests
